Knockout.js! Fixes #28, fixes #3
@gholtslander

Well here it is. Knockout.js functionality for the list page! In addition to that I have added the styles of Bootstrap so we are looking pretty. You might even be able to write posts and stuff from your phone or something now! Haha.

I'll give you a day or so to check this out. I'm super pumped about it though.