Fixes #26 Adding category name as a template field
Since the category name link is part of the template, I figure it makes sense to just expose the category name to the templates and then let the user do whatever they like. I put in an example of how it works. If you have it like this and category name is blank, it'll just show all the posts so it isn't a huge deal.